add pest parser
The grammar is a work in progress. The full MySQL grammar is huge. Since
MySQL is more or less a subset of Postgres (MySQL doesn't support all of
the SQL standard), the plan is to slowly have it successfully parse the
same tests as postgres module first, then add MySQL's peculiar rules.